requirements.txt
coverage
mock
-e git://github.com/lacostej/nosyd.git#egg=nosyd
haigha
pyyaml
